Acomodando footer en index, y he creado helper functions para exportar navbars y footer
