+++ index.html
@@ -47,11 +47,15 @@
             <a class="btn btn-primary btn-lg text-uppercase" href="#" id="boton-portada">Registrarme</a>
             
             <a class="btn btn-secondary btn-lg text-uppercase m-1" href="#" id="boton-ingreso">Ingresar</a>
-            
+            <p class="invisible d-md-block d-none">Lorem ipsum dolor sit amet consectetur adipisicing elit. Quasi corrupti voluptate perspiciatis voluptatem autem quaerat rem corporis, quam delectus laboriosam similique deserunt praesentium doloribus optio voluptatum quisquam veniam hic impedit!
+            Eum, assumenda? Doloremque animi eaque voluptates enim! Rem ad, ipsa laborum assumenda accusantium eius doloribus distinctio illum numquam error ab in, molestiae eum dolor ducimus necessitatibus incidunt. Veritatis, nostrum eius.
+            Quos, earum blanditiis! Molestias qui perspiciatis dolore molestiae distinctio cupiditate ab consequatur doloribus odio ut veritatis minima aperiam quam ea, accusamus modi eaque. Ipsum praesentium placeat amet laborum similique nostrum?
+            Aperiam eum, autem ducimus quae eveniet quod ipsa, fugit architecto laboriosam voluptatum vel nulla quo ullam iure. Modi ducimus ab dignissimos pariatur recusandae at iusto porro reiciendis saepe? Aut, aspernatur!
+            Porro nemo nihil doloribus. Vero delectus soluta ullam illum dignissimos at atque corporis tempora, nihil possimus ex libero labore! Modi, ab quibusdam labore nobis deserunt laudantium eius repellat provident error?</p>
           </div>
         </main>
     <footer></footer>
-    <script src="js/app.js"></script>
+    <script src="js/app.js" type="module"></script>
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.2.2/dist/js/bootstrap.bundle.min.js" integrity="sha384-OERcA2EqjJCMA+/3y+gxIOqMEjwtxJY7qPCqsdltbNJuaOe923+mo//f6V8Qbsw3" crossorigin="anonymous"></script>
 </body>
 </html>
